New `large_tables/0` config (#8)
Add a config option for large_tables/0. Including tables in this list tells the Dumper not to query the total row count or sort by inserted_at. This was the cause of some extremely large tables timing out when rendering the records view.

Additionally, this moves the components into heex files.